Make SoundHandle a RAII wrapper around a RawSoundHandle
Previously we returned an allocated struct from the driver which the driver can change (to mark an unloaded sound) but which is deallocated from the exe
This is a problem with e.g. MSVC which may have separate heaps so all instances allocated from a DLL are deleted when the DLL is unloaded. This results in a dangling pointer.

Solution: Return an opaque DriverHandle (void*) from the DLL which is allocated and freed by the driver only. Then wrap this ASAP into a SoundHandle which calls the driver to free it on destruction of the last SoundHandle reference and also calls a register function for the contained pointer to have the driver reset the driverData pointer once the sound is freed to avoid calling into the driver when it is unloaded already.

Fixes #1280