Refactor AI so AIResourceMaps are always used and calculation uses the same formulas
Basic idea: In the current implementation the value for each point (in the radius where a bld should be built) is always recalculated.
Hence keeping the ResMap up2date is not a goal and in fact incredibly hard: We'd need notifications on when coal, water, ... is used, and recalculate for every building built or destroyed
So don't do that for now.
Instead use the currently used implementation for determining the best position and split into calculation and search. Put that into 2 functions in the resMaps
Write access to the resMaps is now basically impossible hence all updates happen inside the class which makes debugging and testing easier
Single point of truth is CalcResourceValue and GetResourceRating