Cluster mineable resources in random maps
The random map did place resources purely random in a mountain. This is
not great. Manual maps always cluster resources and we should do so too.
However, having all resources in every mountain is also a good thing for
the balance of the game so we want to keep the clusters small. This is a
try to build small clusters while keeping the ratio as it was before.