Fix saving of settings when pressing ESC
When pressing ESC the window was directly deleted instead of first calling Close which triggers a settings update for e.g. the tools window.

Refactor this so deleting a window is purely done by the WindowManager and before deleting Close is always called.
Add tests which catch this.

Also some changes to tests to avoid log spam and fix possibly missed mock-verify calls.