Fix missing productivity of forester
For every building with a production of any kind and hence a productivity the `producedWare` optional must be set.
For special cases like the forester setting it to `GoodType::Nothing` is enough to trigger productivity calculation.

Add test to reproduce current bug (forester productivity is zero) and verify this fixes it.

Fixes #1454