Rename dskHostGame to dskGameLobby
As the screen is shown for all players the name doesn't make sense and is rather confusing.
The "room" where all players assemble for a game is usually called "lobby", however we also have the pre-game lobby for players looking for a game, so use GameLobby to make it clear which one this is.