Adds a cmake option to control if a placeholder is written on install
This option enables the possibility for a package maintainer to set RTTR_GAMEDIR to a directory outside the package itself and do no write operation.
That is us useful for systems where the packaging system prevents write operations outside of the package structure on package build time. A system where this is the case is NixOS.

As the use case is special for some packaging systems and not intended for daily use it is marked as an advanced option.