Do not call GetHome() all dependent_figures
In case a ship just unloaded soldiers at a harbour which is under attack and destroyed,
then there are soldiers that are at the same position as the destroyed harbour
but still on the way to it.

The GoHome method in noFigure::GoHome() does not cover this case.
Since noFigure does not know about the destroyed hardbour, this case
is catched in nobBaseWarehouse::DestroyBuilding()