Avoid accidental copies of GameObjects
As we store GameObject references by pointer we must not copy them.
The only exception is for soldiers where we convert a passive soldier
to an attacker or defender by copying the base attributes.

Mark any copy constructor we don't need as deleted and the others as
explicit and protected.