Fix improved alliances for close start positions
With the addon enabled the territory for the HQ won't be calculated
correctly when they are too close.
The code runs into this condtion that reset the owner for territory
gained by a newly placed HQ:
`ownerOfTriggerBld == oldOwner && oldOwner > 0 && reason == TerritoryChangeReason::Build`
When the HQs are too close the position of the newly placed HQ already
belongs to a player and as ownership wasn't yet set for the position of
the HQ `ownerOfTriggerBld` wrongly refers to that player.

Fix by explicitely handling the case of "building" an HQ by setting
`ownerOfTriggerBld` to the new owner of the point, which is the owner of the HQ.

Fixes #1733