Fix testCheats when running standalone
The test creates a desktop instance that requires access to the
VideoDriver and potentially resources (graphics) which are not available
at this point. This leads to failures when the test is not started after
a test installing a test-videodriver in the same process.

Manually create the required instances without that desktop.